stage 7 done

+++ database/migrations/2024_08_08_185810_create_jobs_table.php
@@ -27,6 +27,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('jobs');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('jobs');
+        }
     }
 };
